+++ utils/http-utils/src/main/scala/pl/touk/nussknacker/http/backend/HttpClientConfig.scala
@@ -2,10 +2,13 @@ package pl.touk.nussknacker.http.backend
 
 import net.ceedubs.ficus.readers.ValueReader
 import org.asynchttpclient.DefaultAsyncHttpClientConfig
+import org.asynchttpclient.filter.{FilterContext, FilterException, RequestFilter}
 import pl.touk.nussknacker.engine.api.process.ProcessName
+import pl.touk.nussknacker.http.backend.DefaultHttpClientConfig.ClientConfigBuilderExtension
 import pl.touk.nussknacker.http.backend.HttpClientConfig.EffectiveHttpClientConfig
 import sttp.client3.SttpBackendOptions
 
+import java.net.InetAddress
 import scala.concurrent.duration._
 
 case class HttpClientConfig(
@@ -16,7 +19,8 @@ case class HttpClientConfig(
     followRedirect: Option[Boolean],
     forceShutdown: Option[Boolean],
     // this can be used to tune single scenario
-    configForProcess: Option[Map[String, HttpClientConfig]]
+    configForProcess: Option[Map[String, HttpClientConfig]],
+    forbiddenHosts: Option[List[String]],
 ) {
 
   def toAsyncHttpClientConfig(processName: Option[ProcessName]): DefaultAsyncHttpClientConfig.Builder = {
@@ -28,6 +32,7 @@ case class HttpClientConfig(
       .setUseNativeTransport(effectiveConfig.useNative)
       .setFollowRedirect(effectiveConfig.followRedirect)
       .setThreadPoolName(processName.map(_.value + s"-http-pool").getOrElse(s"http-pool"))
+      .setForbiddenHostRequestFilter(effectiveConfig.forbiddeHost)
   }
 
   def toSttpBackendOptions(processName: Option[ProcessName]): SttpBackendOptions = {
@@ -47,7 +52,8 @@ case class HttpClientConfig(
       // FIXME: does not work by default?
       useNative = extractConfig(_.useNative, false),
       followRedirect = extractConfig(_.followRedirect, false),
-      forceShutdown = extractConfig(_.forceShutdown, false)
+      forceShutdown = extractConfig(_.forceShutdown, false),
+      forbiddeHost = extractConfig(_.forbiddenHosts, List.empty),
     )
   }
 
@@ -61,7 +67,8 @@ object HttpClientConfig {
       maxPoolSize: Int,
       useNative: Boolean,
       followRedirect: Boolean,
-      forceShutdown: Boolean
+      forceShutdown: Boolean,
+      forbiddeHost: List[String],
   )
 
   // ArbitraryTypeReader cannot handle nested option here... :/
@@ -75,7 +82,8 @@ object HttpClientConfig {
       useNative = forOption[Boolean]("useNative"),
       followRedirect = forOption[Boolean]("followRedirect"),
       forceShutdown = forOption[Boolean]("forceShutdown"),
-      configForProcess = forOption("configForProcess")(mapValueReader(vr))
+      configForProcess = forOption("configForProcess")(mapValueReader(vr)),
+      forbiddenHosts = forOption[List[String]]("forbiddenHosts"),
     )
   })
 
@@ -83,10 +91,40 @@ object HttpClientConfig {
 
 object DefaultHttpClientConfig {
 
-  def apply(): HttpClientConfig = HttpClientConfig(None, None, None, None, None, None, None)
+  def apply(): HttpClientConfig = HttpClientConfig(None, None, None, None, None, None, None, None)
 
   val maxPoolSize: Int = 20
 
   val timeout: FiniteDuration = 10 seconds
+
+  implicit class ClientConfigBuilderExtension(val builder: DefaultAsyncHttpClientConfig.Builder) extends AnyVal {
+
+    def setForbiddenHostRequestFilter(forbiddenHosts: List[String]): DefaultAsyncHttpClientConfig.Builder =
+      if (forbiddenHosts.nonEmpty) {
+        addForbiddenHostsFilter(forbiddenHosts)
+      } else {
+        builder
+      }
+
+    private def addForbiddenHostsFilter(forbiddenHosts: List[String]): DefaultAsyncHttpClientConfig.Builder = {
+      val lowerCaseHosts = forbiddenHosts.map(_.toLowerCase)
+      val lowCaseHostsWithMaybeLoopback = if (lowerCaseHosts.contains("localhost")) {
+        InetAddress.getLoopbackAddress.getHostAddress :: lowerCaseHosts
+      } else {
+        lowerCaseHosts
+      }
+      val hostsAsSet = lowCaseHostsWithMaybeLoopback.toSet
+      builder.addRequestFilter(new RequestFilter {
+        override def filter[T](ctx: FilterContext[T]): FilterContext[T] = {
+          val hostName = ctx.getRequest.getUri.getHost.toLowerCase
+          if (hostsAsSet.contains(hostName)) {
+            throw new FilterException(s"Request to $hostName is forbidden")
+          }
+          ctx
+        }
+      })
+    }
+
+  }
 
 }